Move JS implementation up out of the individual layout types and give it a separate container classname
