+++ packages/block-editor/src/layouts/constrained.js
@@ -22,12 +22,7 @@ import { getCSSRules } from '@wordpress/style-engine';
  * Internal dependencies
  */
 import useSetting from '../components/use-setting';
-import {
-	appendSelectors,
-	getAlignmentsInfo,
-	getBlockGapCSS,
-	getPositionCSS,
-} from './utils';
+import { appendSelectors, getAlignmentsInfo, getBlockGapCSS } from './utils';
 import { getGapCSSValue } from '../hooks/gap';
 import { shouldSkipSerialization } from '../hooks/utils';
 
@@ -234,9 +229,6 @@ export default {
 				blockGapValue
 			);
 		}
-
-		// Add position CSS where applicable.
-		output += getPositionCSS( { selector, style } );
 
 		return output;
 	},
